Split out surrogate definitions into own file.
