+++ manifest.json
@@ -32,6 +32,7 @@
       "lib/jsbn.js",
       "lib/rsa.js",
       "lib/underscore-min.js",
+      "data/surrogates.js",
       "src/surrogates.js",
       "src/multiDomainFirstParties.js",
       "src/incognito.js",
